requirements.txt
# Dependências principais
pandas~=2.3.0
numpy~=2.1.3
joblib~=1.5.1
scikit-learn~=1.7.0
python-dotenv~=1.1.0
keras
python-binance
matplotlib~=3.10.3
requests
jupyterlab

# Windows deps
tf-nightly ;platform_system=="Windows"

# Linux deps
tensorflow~=2.13.0;platform_system=="Unix"
tensorflow[and-cuda];platform_system=="Unix"

tensorflow~=2.19.0
